+++ packages/manager/apps/dedicated/client/app/dedicatedCloud/datacenter/dashboard/dedicatedCloud.datacenter.dashboard.module.js
@@ -3,6 +3,7 @@ import angular from 'angular';
 import datacenterDashboardComponent from '../../../components/dedicated-cloud/datacenter/dashboard';
 import deleteDatacenter from './delete';
 import deleteDrp from './deleteDrp';
+import manageNsxEdges from './manage-nsx-edges';
 import routing from './dedicatedCloud.datacenter.dashboard.routes';
 
 const moduleName = 'ovhManagerDedicatedCloudDatacenterDashboard';
@@ -12,6 +13,7 @@ angular
     datacenterDashboardComponent,
     deleteDatacenter,
     deleteDrp,
+    manageNsxEdges,
   ])
   .config(routing);