+++ packages/manager/apps/web/client/app/hosting/task/hosting-task.constants.js
@@ -1,0 +1,4 @@
+export const HOSTING_TASK_TABLE_ID = 'hosting-tasks-datagrid';
+export default {
+  HOSTING_TASK_TABLE_ID,
+};