+++ packages/manager/core/request-tagger/src/requestTagger.ts
@@ -24,7 +24,11 @@ export interface HeadersOverrides {
   [key: string]: Partial<Headers>;
 }
 
-export const ROUTES_PREFIX = ['/engine/apiv6', '/engine/2api'];
+export const ROUTES_PREFIX = [
+  '/engine/apiv6',
+  '/engine/2api',
+  '/engine/api/v2',
+];
 export const DEFAULT_ROUTES_HEADERS_OVERRIDE: HeadersOverrides = {};
 
 let currentPage = 'bootstrap';
@@ -44,7 +48,8 @@ export const defineApplicationVersion = (version: string): void => {
   window.ovhRequestTaggerApplicationVersion = version;
 };
 
-export const getApplicationVersion = (): string => window.ovhRequestTaggerApplicationVersion;
+export const getApplicationVersion = (): string =>
+  window.ovhRequestTaggerApplicationVersion;
 
 export const generateRequestId = (): string => {
   window.top.ovhRequestTaggerRequestIndex =