+++ packages/manager/core/shell-client/src/hooks/useAuth.ts
@@ -2,7 +2,6 @@ import { useShell } from './useShell';
 
 export const useAuth = () => {
   const shell = useShell();
-
   return shell.auth;
 };