+++ packages/manager/core/shell-client/src/hooks/useUX.ts
@@ -2,7 +2,6 @@ import { useShell } from './useShell';
 
 export const useUX = () => {
   const shell = useShell();
-
   return shell.ux;
 };