+++ packages/manager/modules/pci/src/projects/project/instances/add/add.controller.js
@@ -29,6 +29,8 @@ import {
   LOCAL_ZONE_REGION,
 } from './add.constants';
 
+import { INSTANCE_PRICING_LINKS } from '../instances.constants';
+
 export default class PciInstancesAddController {
   /* @ngInject */
   constructor(
@@ -67,6 +69,9 @@ export default class PciInstancesAddController {
     this.FLOATING_IP_AVAILABILITY_INFO_LINK = FLOATING_IP_AVAILABILITY_INFO_LINK;
     this.LOCAL_ZONE_REGION = LOCAL_ZONE_REGION;
     this.PciProject = PciProject;
+    this.instancePricesLink =
+      INSTANCE_PRICING_LINKS[this.user.ovhSubsidiary] ||
+      INSTANCE_PRICING_LINKS.DEFAULT;
   }
 
   $onInit() {