+++ packages/components/ng-at-internet/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.11.7](https://github.com/ovh/manager/compare/@ovh-ux/ng-at-internet@5.11.6...@ovh-ux/ng-at-internet@5.11.7) (2024-03-13)
+
+**Note:** Version bump only for package @ovh-ux/ng-at-internet
+
+
+
+
+
+## [5.11.6](https://github.com/ovh/manager/compare/@ovh-ux/ng-at-internet@5.11.5...@ovh-ux/ng-at-internet@5.11.6) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/ng-at-internet
+
+
+
+
+
+## [5.11.5](https://github.com/ovh/manager/compare/@ovh-ux/ng-at-internet@5.11.4...@ovh-ux/ng-at-internet@5.11.5) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/ng-at-internet
+
+
+
+
+
+## [5.11.4](https://github.com/ovh/manager/compare/@ovh-ux/ng-at-internet@5.11.3...@ovh-ux/ng-at-internet@5.11.4) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/ng-at-internet
+
+
+
+
+
 ## [5.11.3](https://github.com/ovh/manager/compare/@ovh-ux/ng-at-internet@5.11.2...@ovh-ux/ng-at-internet@5.11.3) (2024-02-21)
 
 **Note:** Version bump only for package @ovh-ux/ng-at-internet