+++ packages/manager-components/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/ovh/manager/compare/@ovhcloud/manager-components@1.4.1...@ovhcloud/manager-components@1.5.0) (2024-03-13)
+
+### Features
+
+- pci vouchers app in reactjs ([d6821ce](https://github.com/ovh/manager/commit/d6821cecd3bde7d884054d8e782e9a1e9dbfddac))
+
+## [1.4.1](https://github.com/ovh/manager/compare/@ovhcloud/manager-components@1.4.0...@ovhcloud/manager-components@1.4.1) (2024-03-07)
+
+### Reverts
+
+- Revert "feat(\*): pci vouchers app in reactjs" ([b62c1a8](https://github.com/ovh/manager/commit/b62c1a8b1cfe63dbc420b660667209c324ffa9ab))
+
+# [1.4.0](https://github.com/ovh/manager/compare/@ovhcloud/manager-components@1.3.0...@ovhcloud/manager-components@1.4.0) (2024-03-07)
+
+### Features
+
+- pci vouchers app in reactjs ([ad08ee4](https://github.com/ovh/manager/commit/ad08ee4618d6243328eee76af7d1bd459a1a7d83))
+
 # [1.3.0](https://github.com/ovh/manager/compare/@ovhcloud/manager-components@1.2.0...@ovhcloud/manager-components@1.3.0) (2024-02-28)
 
 ### Features