+++ packages/manager/apps/hub/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.5.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@5.4.6...@ovh-ux/manager-hub-app@5.5.0) (2024-03-18)
+
+
+### Features
+
+* **hub:** remove IAM and IAM GA banners ([d14804b](https://github.com/ovh/manager/commit/d14804bd5ceb62f377863f736d8d21e5b65c4443))
+
+
+
+
+
+## [5.4.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@5.4.5...@ovh-ux/manager-hub-app@5.4.6) (2024-03-13)
+
+**Note:** Version bump only for package @ovh-ux/manager-hub-app
+
+
+
+
+
+## [5.4.5](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@5.4.4...@ovh-ux/manager-hub-app@5.4.5) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-hub-app
+
+
+
+
+
+## [5.4.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@5.4.3...@ovh-ux/manager-hub-app@5.4.4) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-hub-app
+
+
+
+
+
+## [5.4.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@5.4.2...@ovh-ux/manager-hub-app@5.4.3) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-hub-app
+
+
+
+
+
 ## [5.4.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@5.4.1...@ovh-ux/manager-hub-app@5.4.2) (2024-03-04)
 
 **Note:** Version bump only for package @ovh-ux/manager-hub-app