+++ packages/manager/apps/public-cloud/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.2.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@6.2.7...@ovh-ux/manager-public-cloud@6.2.8) (2024-03-18)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
+## [6.2.7](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@6.2.6...@ovh-ux/manager-public-cloud@6.2.7) (2024-03-13)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
+## [6.2.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@6.2.5...@ovh-ux/manager-public-cloud@6.2.6) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
+## [6.2.5](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@6.2.4...@ovh-ux/manager-public-cloud@6.2.5) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
+## [6.2.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@6.2.3...@ovh-ux/manager-public-cloud@6.2.4) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
 ## [6.2.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@6.2.2...@ovh-ux/manager-public-cloud@6.2.3) (2024-03-04)
 
 **Note:** Version bump only for package @ovh-ux/manager-public-cloud