+++ packages/manager/core/ovh-product-icons/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/ovh/manager/compare/@ovh-ux/ovh-product-icons@0.2.0...@ovh-ux/ovh-product-icons@0.2.1) (2024-03-18)
+
+
+### Bug Fixes
+
+* **container:** safari not rendering svg icons ([15a94f2](https://github.com/ovh/manager/commit/15a94f2a9df02d468dc67594f8880371f64f81d8))
+
+
+
+
+
 # [0.2.0](https://github.com/ovh/manager/compare/@ovh-ux/ovh-product-icons@0.1.0...@ovh-ux/ovh-product-icons@0.2.0) (2024-02-05)