+++ packages/manager/modules/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.16.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-core@12.16.0...@ovh-ux/manager-core@12.16.1) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-core
+
+
+
+
+
 # [12.16.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-core@12.15.3...@ovh-ux/manager-core@12.16.0) (2024-02-21)